+++ lib/iris/_constraints.py
@@ -21,13 +21,10 @@ Provides objects for building up expressions useful for pattern matching.
 import collections
 import operator
 
-import datetime
-import netcdftime
 import numpy as np
 
 import iris.coords
 import iris.exceptions
-from iris.util import _fix_netcdftime_datetime
 
 
 class Constraint(object):
@@ -542,93 +539,4 @@ class TimeConstraint(Constraint):
                 # Scalar time coordinate
                 if n2d(coord.points[0]).hour != self._hour:
                     result_cim.all_false()
-        return result_cim
-
-
-class TimeRangeConstraint(Constraint):
-    """
-    Range selection from a time coordinate.
-
-    Currently, only day-of-year is supported.
-
-    """
-    def __init__(self, day_of_year=None, coord='time'):
-        """
-        Creates a TimeRangeConstraint.
-
-        kwargs:
-
-            * day_of_year - Tuple of (start, end) defining an inclusive range.
-                Both start and end are a (month, day) tuple.
-
-            * coord - The coord to constrain, defaulting to 'time'.
-
-        Example::
-
-            nov_dec = TimeRangeConstraint(day_of_year=((11, 01), (12,31))
-
-        """
-        self.day_of_year, self.coord = day_of_year, coord
-        Constraint.__init__(self)
-
-    def __repr__(self):
-        return 'TimeRangeConstraint(day_of_year={})'.format(self.day_of_year)
-
-    def _day_of_year(self, dt):
-        # May return a netcdftime or datetime object.
-        # Caution: See also iris.util._fix_netcdftime_datetime.
-
-        # netcdf.datetime?
-        if hasattr(dt, 'dayofyr'):
-            result = dt.dayofyr
-        # datetime.datetime
-        else:
-            result = dt.timetuple().tm_yday
-        return result
-
-    def _contains_point(self, point, unit):
-        point_dt = unit.num2date(point)
-        point_yday = self._day_of_year(point_dt)
-
-        start_dt = netcdftime.datetime(point_dt.year, *self.day_of_year[0])
-        start_dt = _fix_netcdftime_datetime(start_dt, unit)
-        start_yday = self._day_of_year(start_dt)
-
-        end_dt = netcdftime.datetime(point_dt.year, *self.day_of_year[1])
-        end_dt = _fix_netcdftime_datetime(end_dt, unit)
-        end_yday = self._day_of_year(end_dt)
-
-        if start_yday < end_yday:
-            result = start_yday <= point_yday <= end_yday
-        elif start_yday > end_yday:
-            result = point_yday >= start_yday or point_yday <= end_yday
-        else:
-            raise ValueError('start_yday == end_yday')
-        return result
-
-    def _CIM_extract(self, cube):
-        result_cim = _ColumnIndexManager(len(cube.shape))
-        try:
-            coord = cube.coord(self.coord)
-        except iris.exceptions.CoordinateNotFoundError:
-            coord = None
-
-        if coord:
-            dims = cube.coord_dims(coord)
-            if len(dims) > 1:
-                raise iris.exceptions.CoordinateMultiDimError(coord)
-
-            # Vector coord?
-            if dims:
-                inout = [self._contains_point(point, coord.units)
-                         for point in coord.points]
-                result_cim[dims[0]] = np.array(inout)
-
-            # Scalar coord.
-            else:
-                if not self._contains_point(coord.points[0], coord.units):
-                    result_cim.all_false()
-        else:
-            result_cim.all_false()
-
         return result_cim
